Create base heatmap renderer
