+++ features/il_reports/lib/src/ui/widgets/asset_heatmap_painter.dart
@@ -1,47 +1,28 @@
 import 'package:flutter/material.dart';
 import 'package:il_core/il_ui_helpers.dart';
 import 'package:il_reports/src/models/asset_heatmap_data.dart';
+import 'package:il_reports/src/ui/helpers/heatmap_renderer.dart';
 
 class AssetHeatmapBackgroundPainter extends CustomPainter {
   final AssetHeatmapData data;
   final IFloorMapImageRenderer imageRenderer;
 
   late FloorMapRenderer floorMapRenderer;
+  late HeatmapRenderer heatmapRenderer;
 
   AssetHeatmapBackgroundPainter({
     required this.data,
     required this.imageRenderer,
   }) {
     floorMapRenderer = FloorMapRenderer();
+    heatmapRenderer = HeatmapRenderer();
   }
 
   @override
   void paint(Canvas canvas, Size size) {
     imageRenderer.draw(canvas);
     floorMapRenderer.drawZones(canvas, data.floorMap, fill: false);
-    _drawHeatmap(canvas);
-  }
-
-  void _drawHeatmap(Canvas canvas) {
-    var paint = Paint();
-    paint.style = PaintingStyle.fill;
-
-    double mapWidth = data.mapSize.width;
-    double mapHeight = data.mapSize.height;
-
-    double cellWidth = data.cellSize.width;
-    double cellHeight = data.cellSize.height;
-
-    int i = 0;
-    for (double y = 0; y < mapHeight; y += cellHeight) {
-      for (double x = 0; x < mapWidth; x += cellWidth) {
-        var cell = data.cells[i];
-        paint.color = data.getCellColor(cell);
-
-        canvas.drawCircle(Offset(x + cellWidth / 2, y + cellHeight / 2), cellWidth / 2 - 4, paint);
-        i++;
-      }
-    }
+    heatmapRenderer.drawHeatmap(canvas, data);
   }
 
   @override
@@ -58,6 +39,7 @@ class AssetHeatmapForegroundPainter extends CustomPainter {
   final AssetHeatmapData data;
 
   late FloorMapRenderer floorMapRenderer;
+  late HeatmapRenderer heatmapRenderer;
 
   AssetHeatmapForegroundPainter({
     required this.transform,
@@ -65,16 +47,13 @@ class AssetHeatmapForegroundPainter extends CustomPainter {
   }) {
     floorMapRenderer = FloorMapRenderer();
     floorMapRenderer.transform = transform;
+
+    heatmapRenderer = HeatmapRenderer();
   }
 
   @override
   void paint(Canvas canvas, Size size) {
     floorMapRenderer.drawZoneLabels(canvas, data.floorMap);
-    _drawLegend(canvas, size);
-  }
-
-  void _drawLegend(Canvas canvas, Size size) {
-    var paint = Paint();
 
     var lPadding = const EdgeInsets.only(top: 20, bottom: 20, right: 20);
     var lSize = Size(10, size.height - lPadding.vertical);
@@ -85,49 +64,7 @@ class AssetHeatmapForegroundPainter extends CustomPainter {
       lSize.height,
     );
 
-    // draw white rect
-
-    paint.style = PaintingStyle.fill;
-    paint.color = const Color.fromARGB(255, 255, 255, 255);
-    canvas.drawRect(lRect, paint);
-
-    // draw gradient
-
-    Shader shader = data.gradient.createShader(lRect);
-    paint.shader = shader;
-
-    canvas.drawRect(lRect, paint);
-    shader.dispose();
-
-    // draw rect border
-
-    paint.color = const Color.fromARGB(255, 0, 0, 0);
-    paint.style = PaintingStyle.stroke;
-    paint.strokeWidth = 2;
-    paint.shader = null;
-
-    canvas.drawRect(lRect, paint);
-
-    // draw labels
-
-    Offset start = Offset(lRect.left - 5, lRect.top + lRect.height);
-
-    for (double p = 0; p <= 1.0; p += 0.2) {
-      Offset lpos = Offset(start.dx, start.dy - p * lRect.height);
-
-      var textPainter = TextPainter(
-        textDirection: TextDirection.ltr,
-        text: TextSpan(
-          text: '${(p * 100).round()}%',
-          style: const TextStyle(color: Color.fromARGB(255, 0, 0, 0), fontSize: 14),
-        ),
-      );
-
-      textPainter.layout();
-
-      Offset tpos = lpos - Offset(textPainter.width, textPainter.height / 2);
-      textPainter.paint(canvas, tpos);
-    }
+    heatmapRenderer.drawLegend(canvas, lRect, data);
   }
 
   @override
